sync Dev to master (#3)
* add template for supporting @event annotation

* finish @message annotation design and sample implementation

* set DefaultEnvironment settings

* feature: add tpl and dist for ask


* enhance: add export the message abi


* define Account and its methods

+++ src/buildins/Block.ts
@@ -3,25 +3,25 @@
  * @author liangqin.fan@gmail.com
  */
 
-import { UInt32 } from "as-scale-codec";
+import { UInt64 } from "as-scale-codec";
 import { ReadBuffer } from "../primitives/readbuffer";
 import { seal_block_number, seal_now } from "../seal/seal0";
 
 export class Block {
-  private _timestamp: UInt32 | null = null;
-  private _number   : UInt32 | null = null;
+  private _timestamp: UInt64 | null = null;
+  private _number   : UInt64 | null = null;
 
-  get timestamp(): u32 {
+  get timestamp(): u64 {
     if (this._timestamp === null) {
-      this._timestamp = ReadBuffer.readInstance<UInt32>(seal_now);
+      this._timestamp = ReadBuffer.readInstance<UInt64>(seal_now);
     }
 
     return this._timestamp!.unwrap();
   }
 
-  get number(): u32 {
+  get number(): u64 {
     if (this._number === null) {
-      this._number = ReadBuffer.readInstance<UInt32>(seal_block_number);
+      this._number = ReadBuffer.readInstance<UInt64>(seal_block_number);
     }
 
     return this._number!.unwrap();
