+++ src/buildins/Callable.ts
@@ -3,17 +3,18 @@
  * @author liangqin.fan@gmail.com
  */
 
-import { UInt32, UInt64 } from "as-scale-codec";
+import { UInt64 } from "as-scale-codec";
 import { ReturnCode } from "../primitives/alias";
 import { ReadBuffer } from "../primitives/readbuffer";
 import { WriteBuffer } from "../primitives/writebuffer";
 import { seal_call } from "../seal/seal0";
-
+import { UInt128 } from "./UInt128";
+import { u128 } from "as-bignum";
 
 export class Callable {
   private _callee: u8[] | null = null;
   private _gas: UInt64 | null = null;
-  private _value: UInt32 | null = null;
+  private _value: UInt128 | null = null;
   private _data: u8[] | null = null;
   private _outBuffer: ReadBuffer | null = null;
 
@@ -26,8 +27,8 @@ export class Callable {
     return this;
   }
 
-  value(v: u32): Callable {
-    this._value = new UInt32(v);
+  value(v: u128): Callable {
+    this._value = new UInt128(v);
     return this;
   }
